requirements.txt
web3==6.11.3
python-dotenv==1.0.0
eth-account==0.10.0
eth-utils==2.3.1
hexbytes==0.3.1
requests==2.31.0